+++ cutplace/test_all.py
@@ -24,7 +24,7 @@ import test_interface
 import test_fields
 import test_ods
 import test_parsers
-import test_range
+import test_ranges
 import test_tools
 import test_web
 import unittest
@@ -36,7 +36,7 @@ import fields
 import interface
 import ods
 import parsers
-import range
+import ranges
 import tools
 import version
 import web
@@ -49,8 +49,8 @@ def createTestSuite():
     loader = unittest.TestLoader()
 
     # TODO: Automatically discover doctest cases.
-    for module in checks, cutplace, data, fields, interface, ods, parsers, range, tools, version, web:
-        result.addTest(doctest.DocTestSuite(module))    
+    for module in checks, cutplace, data, fields, interface, ods, parsers, ranges, tools, version, web:
+        result.addTest(doctest.DocTestSuite(module))
 
     # TODO: Automatically discover test cases.
     allTests = [
@@ -70,15 +70,15 @@ def createTestSuite():
             test_parsers.DelimitedParserTest,
             test_parsers.ExcelReaderTest,
             test_parsers.FixedParserTest,
-            test_range.RangeTest,
+            test_ranges.RangeTest,
             test_tools.ToolsTest
             # FIXME: Stop server and add: test_web.WebTest
             ]
     for testCaseClass in allTests:
         result.addTest(loader.loadTestsFromTestCase(testCaseClass))
-    
+
     return result
-    
+
 def main():
     """
     Run all tests.
@@ -95,7 +95,7 @@ def main():
     print "test_all: ran %d tests with %d failures and %d errors" % (testCount, failureCount, errorCount)
     if (errorCount + failureCount) > 0:
         sys.exit(1)
-    
+
 if __name__ == '__main__': # pragma: no cover
     logging.basicConfig()
     logging.getLogger("cutplace").setLevel(logging.WARNING)